refactor: 목적지 패널의 퍼널 관련 로직을 useDestinationPanelItem 훅으로 분리
