+++ frontend/react-app/src/pages/home/components/panelItem/destination/init/InitPage.tsx
