+++ frontend/react-app/src/pages/home/hooks/destination/useDestinationPanelFunnel.ts
@@ -1,0 +1,42 @@
+import { selectedTravelInfoSelector } from "@/shared/atom/travelAtom";
+import { useFunnel } from "@/shared/hooks/useFunnel";
+import { useSetRecoilState } from "recoil";
+
+type FunnelSteps = "INIT" | "AI_RECOMMENDATION" | "HUMAN_SELECT" | "RESULT";
+
+export default function useDestinationPanelFunnel() {
+  const { step, setStep, Funnel } = useFunnel<FunnelSteps>("INIT");
+  const updateSelectedTravelInfo = useSetRecoilState(selectedTravelInfoSelector);
+
+  const moveToInitialPage = () => {
+    setStep("INIT");
+    updateSelectedTravelInfo((prev) => ({
+      ...prev,
+      destination: {
+        airportCode: "",
+        city: "",
+      },
+    }));
+  };
+
+  const moveToAIRecommendationPage = () => {
+    setStep("AI_RECOMMENDATION");
+  };
+
+  const moveToHumanSelectPage = () => {
+    setStep("HUMAN_SELECT");
+  };
+
+  const moveToResultPage = () => {
+    setStep("RESULT");
+  };
+
+  return {
+    step,
+    Funnel,
+    moveToInitialPage,
+    moveToAIRecommendationPage,
+    moveToHumanSelectPage,
+    moveToResultPage,
+  };
+}
